+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+### Added
+- Added example with slow responding server to examples
+
+
 ## [1.0.6] - 2019-06-07
 ### Added
 - Added support for calendar types `datetime.datetime`, `datetime.time`, `datetime.date`.